@@ -1,11 +1,11 @@
 name: alt_bloc
 description: A library that provides components for more easier and boilerplate implementation of the BLoC architecture pattern.
-version: 1.0.0+3
+version: 2.0.0-nullsafety.0
 
 homepage: https://github.com/dashdevs/alt_bloc
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
   flutter:
